feat: Migrate to @tanstack/react-start with Vite (idiomatic solution)
This is the proper migration from @tanstack/start to the new
@tanstack/react-start package with Vite 7, replacing the previous
hacky pnpm overrides workaround.

Changes:
- Migrate from @tanstack/start to @tanstack/react-start v1.144.0
- Upgrade to Vite 7 (required by react-start)
- Move source from app/ to src/ (Vite convention)
- Replace app.config.ts with vite.config.ts using tanstackStart plugin
- Remove custom client.tsx and ssr.tsx (using default entry points)
- Update router.tsx to export getRouter() instead of createRouter()
- Remove all pnpm overrides from root package.json
- Update CLAUDE.md with new directory structure and idiomatic solutions guideline

The new setup uses:
- @tanstack/react-start: 1.144.0
- @tanstack/react-router: 1.144.0
- @tanstack/react-router-devtools: 1.144.0
- Vite: 7.x
