--- apps/frontend/app/client.tsx
@@ -1,7 +1,0 @@
-import { hydrateRoot } from 'react-dom/client';
-import { StartClient } from '@tanstack/start';
-import { createRouter } from './router';
-
-const router = createRouter();
-
-hydrateRoot(document.getElementById('root')!, <StartClient router={router} />);
