--- apps/frontend/app/router.tsx
@@ -1,15 +1,0 @@
-import { createRouter as createTanStackRouter } from '@tanstack/react-router';
-import { routeTree } from './routeTree.gen';
-
-export function createRouter() {
-  return createTanStackRouter({
-    routeTree,
-    defaultPreload: 'intent',
-  });
-}
-
-declare module '@tanstack/react-router' {
-  interface Register {
-    router: ReturnType<typeof createRouter>;
-  }
-}
