--- apps/frontend/app/ssr.tsx
@@ -1,12 +1,0 @@
-import {
-  createStartHandler,
-  defaultStreamHandler,
-} from '@tanstack/start/server';
-import { getRouterManifest } from '@tanstack/start/router-manifest';
-
-import { createRouter } from './router';
-
-export default createStartHandler({
-  createRouter,
-  getRouterManifest,
-})(defaultStreamHandler);
