+++ apps/frontend/tsconfig.json
@@ -15,10 +15,10 @@
     "noEmit": true,
     "baseUrl": ".",
     "paths": {
-      "~/*": ["./app/*"]
+      "~/*": ["./src/*"]
     },
     "types": ["vite/client"]
   },
-  "include": ["app/**/*", "app.config.ts"],
+  "include": ["src/**/*", "vite.config.ts"],
   "exclude": ["node_modules", ".output", "dist"]
 }
