+++ apps/frontend/vite.config.ts
@@ -1,0 +1,15 @@
+import { defineConfig } from 'vite';
+import { tanstackStart } from '@tanstack/react-start/plugin/vite';
+import tsConfigPaths from 'vite-tsconfig-paths';
+
+export default defineConfig({
+  server: {
+    port: 3000,
+  },
+  plugins: [
+    tanstackStart(),
+    tsConfigPaths({
+      projects: ['./tsconfig.json'],
+    }),
+  ],
+});
